requirements.txt
PyQt6
pygame
flask
flask-login
werkzeug
waitress
requests
cryptography
argon2-cffi
mutagen
flask-limiter
